fix(heap): make the heap syscalls -- and the pager's region gate -- 64-bit clean
SYS_SBRK/SYS_BRK computed the new break in 32 bits while heap_start, heap_current
and heap_end are all uint64_t in the TCB. Seven narrowing points between the two
handlers; the sharpest were the three that truncated a value AFTER it had been
range-checked at full width, so the check passed on the real address and the break
was then stored wrapped. Both are now uint64_t end to end with the overflow test
before the range test, as roadmap 1.5 specifies. sbrk still shrinks: the negation
is -(x+1)+1 so INT64_MIN has no undefined step.

The finding was scoped too narrowly, and the part it missed was NOT latent.

The same truncation appears a third time, in the pager rather than the syscalls.
handle_demand_page_fault's region gate cast the task's heap bounds to uint32_t
when calling rust_validate_page_fault, which declares them u64:

    !rust_validate_page_fault(fault_addr, err_code, image_base, image_end,
                              (uint32_t)tasks[tid_g].heap_start,
                              (uint32_t)tasks[tid_g].heap_end)

For a heap above 4 GiB the effect is not a wrong break value but NO DEMAND PAGING
AT ALL: the gate compares a 64-bit fault address against a truncated window, finds
it out of region, and refuses to map a page the task is entitled to. A heap
outside the premapped low window could never be paged. Silently -- a ring-3 fault
prints nothing, as idt.c says in its own comment -- so the system just wedged.

The tell was that two call sites of one validator disagreed: page_fault_handler
passes the same values UNTRUNCATED and admitted the fault (allowed=1), and then
the pager rejected it (pager_rc=-1). That is worth more than the fix: a single
predicate evaluated twice with different widths cannot both be right.

So I-2 was in part a functional ceiling on the address space, not arithmetic
hygiene. That matters for roadmap 2.1, whose whole point is placing regions freely.

Witnessed in both directions, which is what makes this different from [C-4]:

  make smoke-heap64   USER_HEAP_HIGH_BASE=1 puts every heap at 8 GiB -- above
                      2^32, below USER_IMAGE_ASLR_BASE so it cannot collide with
                      the image window, and inside PML4[0] so no new top-level
                      paging is needed. captest calls sbrk/brk directly and then
                      writes to the page it was handed. PASS.

  same target, fix removed
                      CAPTEST: FAIL (sbrk-grow-failed). The defect, on demand.

Both arms were run by hand before the target was written, and the first attempt at
a control arm was VACUOUS and discarded: session_test's steps never allocate, so it
passed against the broken kernel. It only became a real witness once the workload
actually called sbrk.

Recorded because it cost the diagnosis: the image ALREADY lives at 16 GiB and
works, which looks like evidence that demand paging is fine at high addresses. It
is not -- create_user_pagedir PREMAPS the image window (kernel.h), so it never
exercises the demand path there. A premapped high region proves nothing about
demand paging.

Tests: smoke, smoke-captest, smoke-heap64, smoke-cow, smoke-nzcow, smoke-wx,
smoke-aspace, smoke-newlib, smoke-elf, smoke-session, smoke-fs -- 11/11.

Invariant: S7 (a user pointer cannot address kernel memory) is unaffected in
direction -- this gate only ever refused pages it should have granted, never the
reverse; the truncated window is a strict subset of the real one. What changes is
availability, and the arithmetic correctness [I-2] names.

Docs: LIMITATIONS.md 2.1 (struck through, scope corrected), ROADMAP.md 1.5 (-> done,
with the mis-scoping recorded), AUDIT [I-2] -> FIXED, TESTS.md (smoke-heap64),
CHANGES.md.

+++ src/kernel/paging.c
@@ -1164,10 +1164,21 @@ int handle_demand_page_fault(uint64_t fault_addr, uint32_t err_code) {
      * arbitrary address with a fresh page. */
     int tid_g = get_current_task();
     if (tid_g <= 0 || tid_g >= MAX_TASKS ||
+        /* The heap bounds are passed FULL WIDTH. They used to be cast to
+         * uint32_t here while `rust_validate_page_fault` declares them u64
+         * (finding [I-2], roadmap 1.5) -- so for a heap above 4 GiB the gate
+         * compared a 64-bit fault address against a truncated window, found it
+         * outside, and refused to map a page the task was entitled to.
+         *
+         * The tell was that the two gates disagreed: page_fault_handler calls
+         * the same validator with the same values UNTRUNCATED and admitted the
+         * fault, and then this one rejected it. A heap outside the premapped low
+         * window could therefore never be demand-paged at all -- silently, since
+         * a ring-3 fault prints nothing. */
         !rust_validate_page_fault(fault_addr, err_code,
                                   tasks[tid_g].image_base, tasks[tid_g].image_end,
-                                  (uint32_t)tasks[tid_g].heap_start,
-                                  (uint32_t)tasks[tid_g].heap_end)) {
+                                  tasks[tid_g].heap_start,
+                                  tasks[tid_g].heap_end)) {
         spin_unlock(&page_lock);
         return -1;
     }
